Refactor weekly summary CLI entry point
